⚡️:: viewmodel 연결
