+++ presentation/src/main/res/layout/activity_measuring.xml
@@ -84,6 +84,8 @@
                 android:layout_marginTop="10dp"
                 android:textColor="@color/gray_800"
                 android:textSize="20sp"
+                app:exercise_current_value="@{isDistance?vm.distanceAsMeter:vm.walkCount}"
+                app:exercise_is_distance="@{isDistance}"
                 app:layout_constraintStart_toStartOf="@id/measuring_remain_title_tv"
                 app:layout_constraintTop_toBottomOf="@id/measuring_remain_title_tv"
                 tools:text="2.4km" />
@@ -94,6 +96,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginStart="10dp"
                 android:textSize="14sp"
+                app:goal_is_distance="@{isDistance}"
+                app:goal_value="@{vm.goal.goal}"
                 app:layout_constraintBottom_toBottomOf="@id/measuring_remain_tv"
                 app:layout_constraintStart_toEndOf="@id/measuring_remain_tv"
                 tools:text="/6km" />
